requirements.txt
calplot==0.1.7.5
Flask_Login==0.6.2
flask_sqlalchemy==3.0.3
Flask_WTF==1.0.1
matplotlib==3.5.1
pandas==1.5.0
plotnine==0.10.1
SQLAlchemy==1.4.44
WTForms==3.0.1
WTForms_Components==0.10.5
